\name{mCSEAdata-package}
\alias{mCSEAdata-package}
\alias{mCSEAdata}
\docType{package}
\title{
    Data and examples for mCSEA package
}
\description{
    Association files between Illumina's 450K and EPIC microarrays
    probes and promoters, CpG Islands and gene bodies. Annotation for the
    previous microarrays' probes. There are also
    sample data from EPIC platform and expression microarrays to test mCSEA 
    functions and a necessary object for mCSEAPlot function.
}
\author{
Jordi Martorell Marugán

Maintainer: Jordi Martorell Marugán<jordi.martorell@genyo.es>
}

\examples{
data(mcseadata)
data(bandTable)
}

\keyword{datasets}
